+++ src/lib/components/table/proposals/ProposalsTableRow.tsx
@@ -6,12 +6,12 @@ import { useTrack } from "lib/amplitude";
 import { useBaseApiRoute, useCelatoneApp } from "lib/app-provider";
 import { ExplorerLink, getNavigationUrl } from "lib/components/ExplorerLink";
 import { StopPropagationBox } from "lib/components/StopPropagationBox";
-import { Proposer } from "lib/components/table/proposals/Proposer";
 import type { Option, Proposal } from "lib/types";
 import { ProposalStatus } from "lib/types";
 import { openNewTab } from "lib/utils";
 
 import { ProposalTextCell } from "./ProposalTextCell";
+import { Proposer } from "./Proposer";
 import { ResolvedHeight } from "./ResolvedHeight";
 import { StatusChip } from "./StatusChip";
 import { VotingEndTime } from "./VotingEndTime";
