+++ src/lib/components/table/proposals/index.ts
@@ -1,1 +1,2 @@
 export * from "./ProposalsTable";
+export * from "./StatusChip";
