+++ src/lib/pages/contract-details/components/tables/migration/index.tsx
@@ -6,11 +6,11 @@ import { Pagination } from "lib/components/pagination";
 import { usePaginator } from "lib/components/pagination/usePaginator";
 import { EmptyState } from "lib/components/state";
 import { TableContainer } from "lib/components/table";
-import { MigrationMobileCard } from "lib/pages/contract-details/components/tables/migration/MigrationMobileCard";
 import { useMigrationHistories } from "lib/pages/contract-details/data";
 import type { ContractAddr, Option } from "lib/types";
 
 import { MigrationHeader } from "./MigrationHeader";
+import { MigrationMobileCard } from "./MigrationMobileCard";
 import { MigrationRow } from "./MigrationRow";
 
 interface MigrationTableProps {
