+++ src/lib/pages/proposals/components/ProposalStatusFilter.tsx
@@ -5,11 +5,13 @@ import type { Dispatch, SetStateAction } from "react";
 import { useMemo, useState, useRef, forwardRef } from "react";
 
 import { AmpEvent, useTrack } from "lib/amplitude";
-import { FilterChip } from "lib/components/filter/FilterChip";
-import { DropdownContainer } from "lib/components/filter/FilterComponents";
-import { FilterDropdownItem } from "lib/components/filter/FilterDropdownItem";
-import { FilterInput } from "lib/components/filter/FilterInput";
-import { StatusChip } from "lib/components/table/proposals/StatusChip";
+import {
+  FilterChip,
+  DropdownContainer,
+  FilterDropdownItem,
+  FilterInput,
+} from "lib/components/filter";
+import { StatusChip } from "lib/components/table";
 import { ProposalStatus } from "lib/types";
 
 export interface ProposalStatusFilterProps extends InputProps {
