+++ src/lib/pages/proposals/components/ProposalTypeFilter.tsx
@@ -11,10 +11,12 @@ import type { Dispatch, SetStateAction } from "react";
 import { useMemo, useState, useRef, forwardRef } from "react";
 
 import { AmpEvent, useTrack } from "lib/amplitude";
-import { FilterChip } from "lib/components/filter/FilterChip";
-import { DropdownContainer } from "lib/components/filter/FilterComponents";
-import { FilterDropdownItem } from "lib/components/filter/FilterDropdownItem";
-import { FilterInput } from "lib/components/filter/FilterInput";
+import {
+  FilterChip,
+  DropdownContainer,
+  FilterDropdownItem,
+  FilterInput,
+} from "lib/components/filter";
 import { CustomIcon } from "lib/components/icon";
 import { useProposalTypes } from "lib/services/proposalService";
 import type { ProposalType } from "lib/types";
