+++ src/lib/pages/public-project/components/tables/account/PublicProjectAccountRow.tsx
@@ -1,34 +1,16 @@
 import { Grid, Text } from "@chakra-ui/react";
 
-import type { NavigationArgs } from "lib/app-provider";
 import { useInternalNavigate } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { TableRow } from "lib/components/table";
 import type { Account } from "lib/types";
 
+import { getNavigationArgs } from "./utils";
+
 interface AccountTableRowProps {
   accountInfo: Account;
   templateColumns: string;
 }
-
-export const getNavigationArgs = (accountInfo: Account): NavigationArgs => {
-  switch (accountInfo.type) {
-    case "account":
-      return {
-        pathname: "/accounts/[accountAddress]",
-        query: { accountAddress: accountInfo.address },
-      };
-    case "contract":
-      return {
-        pathname: "/contracts/[accountAddress]",
-        query: { accountAddress: accountInfo.address },
-      };
-    default:
-      return {
-        pathname: "",
-      };
-  }
-};
 
 export const PublicProjectAccountRow = ({
   accountInfo,
