+++ src/lib/pages/public-project/components/tables/code/PublicProjectCodeMobileCard.tsx
@@ -3,16 +3,16 @@ import { Flex, Text } from "@chakra-ui/react";
 import { useInternalNavigate } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { PermissionChip } from "lib/components/PermissionChip";
-import { MobileCardTemplate } from "lib/components/table";
-import { MobileLabel } from "lib/components/table/MobileLabel";
-import type { PublicCode } from "lib/types";
+import { MobileCardTemplate, MobileLabel } from "lib/components/table";
 import { getCw2Info } from "lib/utils";
 
+import type { PublicCodeInfo } from ".";
+
 interface PublicProjectCodeMobileCardProps {
-  publicInfo: PublicCode;
+  publicCodeInfo: PublicCodeInfo;
 }
 export const PublicProjectCodeMobileCard = ({
-  publicInfo,
+  publicCodeInfo: { publicInfo },
 }: PublicProjectCodeMobileCardProps) => {
   const cw2Info = getCw2Info(publicInfo.cw2Contract, publicInfo.cw2Version);
   const navigate = useInternalNavigate();
