+++ src/lib/pages/public-project/components/tables/contract/index.tsx
@@ -1,4 +1,4 @@
-import { TableContainer, Grid, Box, Flex } from "@chakra-ui/react";
+import { Grid, Box, Flex } from "@chakra-ui/react";
 import { matchSorter } from "match-sorter";
 import { observer } from "mobx-react-lite";
 import { useMemo, useState } from "react";
@@ -6,7 +6,12 @@ import { useMemo, useState } from "react";
 import { useMobile } from "lib/app-provider";
 import { TextInput } from "lib/components/forms";
 import { EmptyState } from "lib/components/state";
-import { TableHeader, TableTitle, ViewMore } from "lib/components/table";
+import {
+  TableContainer,
+  TableHeader,
+  TableTitle,
+  ViewMore,
+} from "lib/components/table";
 import { useContractStore } from "lib/providers/store";
 import type { ContractLocalInfo } from "lib/stores/contract";
 import type { PublicContract, Option, ContractAddr } from "lib/types";
